refactor: group commands together
Remove a bit of code complexity by grouping all the commands together.
