+++ src/command/init.rs
@@ -1,11 +1,10 @@
-use crate::emoji;
 use crate::CLOG;
 use anyhow::Context;
 use console::style;
 use std::fs;
 use std::path::PathBuf;
 
-pub fn init_prjfmt(path: Option<PathBuf>) -> anyhow::Result<()> {
+pub fn init_cmd(path: Option<PathBuf>) -> anyhow::Result<()> {
     let file = match path {
         Some(loc) => loc,
         None => PathBuf::from("."),
@@ -15,7 +14,7 @@ pub fn init_prjfmt(path: Option<PathBuf>) -> anyhow::Result<()> {
     fs::write(
         file_path.as_path(),
         r#"# prjfmt is the universal code formatter - https://github.com/numtide/prjfmt
-[formats.<Language>]
+[formatter.<Language>]
 includes = [ "*.<language-extension>" ]
 excludes = []
 command = ""
@@ -24,8 +23,7 @@ options = []
     )
     .with_context(|| {
         format!(
-            "{} {} `{}`",
-            emoji::ERROR,
+            "{} `{}`",
             style("Error writing").bold().red(),
             style(file_path.display()).bold()
         )
