+++ src/command/mod.rs
@@ -1,9 +1,12 @@
 //! CLI command structures for formatter
 #![allow(clippy::redundant_closure)]
 
+mod format;
 mod init;
 
-use self::init::init_prjfmt;
+use self::format::format_cmd;
+use self::init::init_cmd;
+use crate::customlog::LogLevel;
 use std::path::PathBuf;
 use structopt::StructOpt;
 
@@ -18,9 +21,39 @@ pub enum Command {
     },
 }
 
-/// Run a command with the given logger!
-pub fn run_prjfmt_cli(command: Command) -> anyhow::Result<()> {
-    match command {
-        Command::Init { path } => init_prjfmt(path),
+/// ✨  format all your language!
+#[derive(Debug, StructOpt)]
+pub struct Cli {
+    /// Main command to run
+    #[structopt(subcommand)]
+    pub cmd: Option<Command>,
+
+    /// Log verbosity is based off the number of v used
+    #[structopt(long = "verbose", short = "v", parse(from_occurrences))]
+    pub verbosity: u8,
+
+    #[structopt(long = "quiet", short = "q")]
+    /// No output printed to stdout
+    pub quiet: bool,
+
+    #[structopt(long = "log-level", default_value = "debug")]
+    /// The maximum level of messages that should be logged by prjfmt. [possible values: info, warn, error]
+    pub log_level: LogLevel,
+
+    /// Files to format
+    pub files: Option<PathBuf>,
+}
+
+/// Run a command with the given logger
+pub fn run_cli(cli: Cli) -> anyhow::Result<()> {
+    if let Some(command) = cli.cmd {
+        match command {
+            Command::Init { path } => init_cmd(path)?,
+        }
+    } else {
+        format_cmd(cli)?;
+        return Ok(());
     }
+
+    return Ok(());
 }
