+++ src/main.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::redundant_closure, clippy::redundant_pattern_matching)]
 
-use prjfmt::{run_cli, Cli, CLOG};
+use prjfmt::command::{run_cli, Cli};
+use prjfmt::CLOG;
 use structopt::StructOpt;
 
 fn main() {
